Move `perform_full_sync_starting_from` to `sync`
Keep to closer to the synchronization utility functions.
